+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 ### Added
 - Validate structure of Akoya CODEX submissions.
 - Generate submission template.
+- Check that fixture-based tests actually ran.
 - Check types early, rather than waiting for file-not-found.